requirements.txt
flask
flask-socketio
pyttsx3
openai-whisper
whisperx
torch